Move tests folder out of src
