+++ tests/AppRouter.test.jsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 import { MemoryRouter } from 'react-router-dom';
-import AppRouter from '../AppRouter';
+import AppRouter from '../src/AppRouter';
 
-vi.mock('../App.jsx', () => ({ default: () => <div>Mock FSH Online</div> }));
+vi.mock('../src/App.jsx', () => ({ default: () => <div>Mock FSH Online</div> }));
 
 test('Renders FSH Online App when visiting /FSHOnline', () => {
   const { getByText } = render(
