+++ vite.config.mjs
@@ -59,6 +59,6 @@ export default defineConfig({
   test: {
     environment: 'jsdom',
     globals: true,
-    setupFiles: 'src/setupTests.js'
+    setupFiles: 'tests/setupTests.js'
   }
 });
